Relieve build.gradle from backend-specific tasks (#97)
* Explicit JPA Metamodel generation (no more generatedJpaClasses for
   modules that don't actually need meta classes)
* Overall Gradle cleanup
