+++ backend-boarding/build.gradle
@@ -1,3 +1,5 @@
+apply from: "$rootDir/env-backend.gradle"
+
 dependencies {
     compile project(":backend-shared")
 
@@ -8,3 +10,5 @@ dependencies {
     compile libraries.application.springBootWeb
     compile libraries.application.springDataRest
 }
+
+bootRepackage.enabled = false
