+++ backend-shared/build.gradle
@@ -1,15 +1,19 @@
+apply from: "$rootDir/env-backend.gradle"
+apply from: "$rootDir/ext-jpametamodel.gradle"
+
 dependencies {
     compile libraries.commons.moneta
     compile libraries.commons.apacheLang3
     compile libraries.commons.apacheValidator
 
     compile libraries.persistence.jpaApi
-    compile libraries.persistence.springData
+    compile libraries.persistence.springDataJpa
 
-    compile (libraries.bus.embeddedKafka) {exclude group: 'org.slf4j'}
+    compile (libraries.bus.embeddedKafka) {exclude module: 'log4j'}
     compile libraries.bus.springCloudStreamKafka
-
     compile libraries.application.springDataRest
 
-  testCompile libraries.test.spring
+    testCompile libraries.test.spring.springCloudStreamKafka
 }
+
+bootRepackage.enabled = false
