+++ backend-shared/src/main/java/eu/socialedge/vega/backend/infrastructure/eventbus/cloudstream/DomainEventSource.java
@@ -12,19 +12,16 @@
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
  */
-package eu.socialedge.vega.backend.infrastructure.eventbus.kafka;
+package eu.socialedge.vega.backend.infrastructure.eventbus.cloudstream;
 
-import eu.socialedge.vega.backend.ddd.DomainEventPublisher;
-import eu.socialedge.vega.backend.ddd.DomainEventRegistrar;
+import org.springframework.cloud.stream.annotation.Output;
+import org.springframework.messaging.MessageChannel;
 
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
+public interface DomainEventSource {
 
-/**
- * This configuration enables Spring Cloud Stream based implementations
- * of {@link DomainEventPublisher} and {@link DomainEventRegistrar}.
- */
-@Configuration
-@ComponentScan("eu.socialedge.vega.backend.infrastructure.eventbus")
-public class SpringEventBusConfig {
+    String CHANNEL_NAME = "eventBus";
+
+    @Output(DomainEventSource.CHANNEL_NAME)
+    MessageChannel eventBus();
+
 }
