+++ backend-shared/src/main/java/eu/socialedge/vega/backend/infrastructure/eventbus/cloudstream/SpringDomainEventPublisher.java
@@ -12,11 +12,10 @@
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
  */
-package eu.socialedge.vega.backend.infrastructure.eventbus;
+package eu.socialedge.vega.backend.infrastructure.eventbus.cloudstream;
 
 import eu.socialedge.vega.backend.ddd.DomainEvent;
 import eu.socialedge.vega.backend.ddd.DomainEventPublisher;
-import eu.socialedge.vega.backend.infrastructure.eventbus.kafka.DomainEventSource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
