+++ backend-terminal/build.gradle
@@ -1,12 +1,14 @@
+apply from: "$rootDir/env-backend.gradle"
+
 dependencies {
-  compile project(":backend-shared")
+    compile project(":backend-shared")
 
-  compile libraries.persistence.hibernateCore
-  compile libraries.persistence.postgresDriver
-  compile libraries.persistence.h2Driver
+    compile libraries.persistence.hibernateCore
+    compile libraries.persistence.postgresDriver
+    compile libraries.persistence.h2Driver
 
-  compile libraries.application.springBoot
-  compile libraries.application.springBootData
-  compile libraries.application.springBootWeb
-  compile libraries.application.springDataRest
+    compile libraries.application.springBoot
+    compile libraries.application.springBootData
+    compile libraries.application.springBootWeb
+    compile libraries.application.springDataRest
 }
