+++ ext-jpametamodel.gradle
@@ -1,0 +1,30 @@
+configurations {
+  jpaMetamodel
+}
+
+dependencies {
+  jpaMetamodel libraries.gen.jpaModel
+}
+
+sourceSets {
+  generated {
+    java {
+      srcDirs = ['src/main/generated']
+    }
+  }
+}
+
+sourceSets.main.java.srcDirs += sourceSets.generated.java.srcDirs.iterator().next()
+
+task generatedJpaClasses(type: JavaCompile, group: 'build', description: 'metamodel generate') {
+  source = sourceSets.main.java
+  classpath = sourceSets.main.compileClasspath + configurations.jpaMetamodel
+  options.compilerArgs = ["-proc:only"]
+  destinationDir = sourceSets.generated.java.srcDirs.iterator().next()
+}
+
+compileJava.dependsOn generatedJpaClasses
+
+clean {
+  delete sourceSets.generated.java.srcDirs
+}
